// src/types/partials/BundleEntry.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TBundleLink } from '../partials/BundleLink';
import { TUri } from '../simpleTypes/Uri';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TBundleSearch } from '../partials/BundleSearch';
import { TBundleRequest } from '../partials/BundleRequest';
import { TBundleResponse } from '../partials/BundleResponse';

export type TBundleEntry = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    link?: TBundleLink[];
    fullUrl?: TUri;
    resource?: TResourceContainer;
    search?: TBundleSearch;
    request?: TBundleRequest;
    response?: TBundleResponse;
};